requirements.txt
beautifulsoup4==4.12.2
discord.py==2.3.1
pip==22.1.2
requests==2.27.1
